Add symfony validation

+++ src/Controller/RecipesController.php
@@ -2,7 +2,11 @@
 
 namespace App\Controller;
 
+use App\CategorySchema;
+use App\Entity\CategoriesEntity;
 use App\Entity\RecipesEntity;
+use App\RecipesPostSchema;
+use App\RecipesUpdateSchema;
 use App\Repository\RecipesRepository;
 use App\RecipesLogger;
 use Monolog\Logger;
@@ -14,6 +18,9 @@ use Exception;
 use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Component\Validator\Constraints as Assert;
 /**
  * Class RecipesController
  *
@@ -35,12 +42,7 @@ class RecipesController extends AbstractController {
     const CATEGORY_DESSERT = 'dessert';
     const CATEGORY_SALAD = 'salad';
     const CATEGORY_APPETIZER = 'appetizer';
-//    const CATEGORY_MAIN_DISH_BEEF = 'main dish beef';
-//    const CATEGORY_MAIN_DISH_POULTRY = 'main dish poultry';
-//    const CATEGORY_MAIN_DISH_PORK = 'main dish pork';
-//    const CATEGORY_MAIN_DISH_SEAFOOD = 'main dish seafood';
-//    const CATEGORY_MAIN_DISH_VEGETARIAN = 'main dish vegetarian';
-   const CATEGORY_MAIN_DISH = 'main dish';
+    const CATEGORY_MAIN_DISH = 'main dish';
     const CATEGORY_HOLIDAYS = 'holidays';
     const CATEGORY_BREAKFAST = 'breakfast';
     const CATEGORY_SIDE_DISH = 'side dish';
@@ -59,11 +61,6 @@ class RecipesController extends AbstractController {
         self::CATEGORY_SALAD,
         self::CATEGORY_SIDE_DISH,
         self::CATEGORY_SOUP
-//        self::CATEGORY_MAIN_DISH_POULTRY,
-//        self::CATEGORY_MAIN_DISH_BEEF,
-//        self::CATEGORY_MAIN_DISH_PORK,
-//        self::CATEGORY_MAIN_DISH_SEAFOOD,
-//        self::CATEGORY_MAIN_DISH_VEGETARIAN
         ];
 
     /** @var RecipesLogger  */
@@ -84,7 +81,7 @@ class RecipesController extends AbstractController {
     /**
      * SensorController constructor.
      *
-     * @param RecipesRepository|null $recipesRepository
+     * @param RecipesRepository $recipesRepository
      * @param RecipesLogger $logger
      * @param ObjectNormalizer $objectNormalizer
      */
@@ -100,7 +97,6 @@ class RecipesController extends AbstractController {
         $this->serializer = new Serializer($normalizers, $encoders);
         $this->response->headers->set('Content-Type', 'application/json');
         $this->time_start = microtime(true);
-
         $this->request  = new Request();
         $this->logger = $logger;
         $this->recipesRepository = $recipesRepository;
@@ -130,10 +126,29 @@ class RecipesController extends AbstractController {
      */
     public function getByName(string $name): Response {
         $name = strtolower($name);
-        $results = $this->recipesRepository->findByQuery(['name' => $name]);
-        $this->normalize($results);
-        $this->validateResponse($results, $name);
+        $valid = !empty($name);
+        if ($valid) {
+            $results = $this->recipesRepository->findByQuery(['name' => $name]);
+            $this->normalize($results);
+            $this->validateResponse($results, $name);
+        }
         $this->updateResponseHeader();
+        return $this->response;
+    }
+
+    /**
+     * Post a category.
+     *
+     * @Route("recipes/category",  methods={"POST", "OPTIONS"}, name="post_category")
+     * @param Request $request
+     * @return Response
+     * @throws Exception
+     */
+    public function postCategory(Request $request, ValidatorInterface $validator, CategorySchema $categorySchema): Response {
+        $input = (array)json_decode($request->getContent());
+        $violations = $validator->validate($input, $categorySchema::$schema);
+        // TODO Category Post
+        $this->validateRequest($violations);
         return $this->response;
     }
 
@@ -171,12 +186,15 @@ class RecipesController extends AbstractController {
      * @throws \Doctrine\ORM\ORMException
      * @throws \Doctrine\ORM\OptimisticLockException
      */
-    public function delete($id) {
-        $result = $this->recipesRepository->delete($id);
-        if (!$result) {
-            $this->response->setStatusCode(self::STATUS_NOT_FOUND);
-        } else {
-            $this->response->setStatusCode(self::STATUS_OK);
+    public function delete($id): Response {
+        if (is_numeric($id)) {
+            $result = $this->recipesRepository->delete($id);
+            if (!$result) {
+                $this->response->setStatusCode(self::STATUS_NOT_FOUND);
+                $this->response->setContent(self::VALIDATION_NO_RECORD);
+            } else {
+                $this->response->setStatusCode(self::STATUS_OK);
+            }
         }
         $this->updateResponseHeader();
         return $this->response;
@@ -186,52 +204,50 @@ class RecipesController extends AbstractController {
     /**
      * Update a recipe.
      *
-     * @Route("recipes/update/{id}", methods={"PUT"}, name="update_recipe")
+     * @Route("recipes/update/{id}", methods={"PATCH"}, name="update_recipe")
      * @param $id
      * @param Request $request
      * @return Response
      * @throws \Doctrine\ORM\ORMException
      * @throws \Doctrine\ORM\OptimisticLockException
      */
-    public function put($id, Request $request): Response {
-        // TODO Validate.
-        // Update Fields.
-        //name, prep_time, cooking_time, category, directions, ingredients, favourites, calories, cuisine, url
-
-        $recipe = $this->recipesRepository->findOneBy(['id' => $id]);
-        if (!empty($recipe)) {
-            $data = json_decode($request->getContent(), true);
-            $data = $this->normalizeData($data);
-            empty($data['name']) ? true : $recipe->setName($data['name']);
-            empty($data['prep_time']) ? true : $recipe->setPrepTime($data['prep_time']);
-            empty($data['cooking_time']) ? true : $recipe->setCookingTime($data['cooking_time']);
-            empty($data['category']) ? true : $recipe->setCategory($data['category']);
-            empty($data['directions']) ? true : $recipe->setDirections($data['directions']);
-            empty($data['ingredients']) ? true : $recipe->setIngredients($data['ingredients']);
-            empty($data['favourites']) ? true : $recipe->setFavourites($data['favourites']);
-            empty($data['calories']) ? true : $recipe->setCalories($data['calories']);
-            empty($data['cuisine']) ? true : $recipe->setCookingTime($data['cooking_time']);
-            empty($data['url']) ? true : $recipe->setCookingTime($data['url']);
-            $validFields = $this->validateRecipeFields($data, 'POST');
-            $valid = true;
-            if (!empty($data['category'])) {
-                $valid = $this->validateCategory($data['category']);
-            }
-            if ($valid && $validFields) {
-                $updatedRecipe = $this->recipesRepository->updateRecipe($recipe);
-                if ($updatedRecipe instanceof RecipesEntity) {
-                    $this->getByIdInternal($id);
+    public function patch($id, Request $request, ValidatorInterface $validator, RecipesUpdateSchema $recipesUpdateSchema): Response {
+        $data = json_decode($request->getContent(), true);
+        $violations = $validator->validate($data, $recipesUpdateSchema::$schema);
+        $valid = $this->validateRequest($violations);
+        if ($valid) {
+            $recipe = $this->recipesRepository->findOneBy(['id' => $id]);
+            if (!empty($recipe)) {
+                // TODO Refactor this nonsense.
+                empty($data['name']) ? true : $recipe->setName($data['name']);
+                empty($data['prepTime']) ? true : $recipe->setPrepTime($data['prepTime']);
+                empty($data['cookingTime']) ? true : $recipe->setCookingTime($data['cookingTime']);
+                empty($data['category']) ? true : $recipe->setCategory($data['category']);
+                empty($data['directions']) ? true : $recipe->setDirections($data['directions']);
+                empty($data['ingredients']) ? true : $recipe->setIngredients($data['ingredients']);
+                empty($data['favourites']) ? true : $recipe->setFavourites($data['favourites']);
+                empty($data['calories']) ? true : $recipe->setCalories($data['calories']);
+                empty($data['cuisine']) ? true : $recipe->setCuisine($data['cuisine']);
+                empty($data['addedBy']) ? true : $recipe->setAddedBy($data['addedBy']);
+                empty($data['url']) ? true : $recipe->setUrl($data['url']);
+                if (!empty($data['category'])) {
+                    $valid = $this->validateCategory($data['category']);
+                }
+                if ($valid) {
+                    $updatedRecipe = $this->recipesRepository->updateRecipe($recipe);
+                    if (!empty($updatedRecipe)) {
+                        $this->getByIdInternal($id);
+                    } else {
+                        $this->response->setStatusCode(self::STATUS_NO_CONTENT);
+                        $this->logger->log(self::VALIDATION_FAILED, ['fields' => $updatedRecipe], Logger::ALERT);
+                    }
                 } else {
-                    $this->response->setStatusCode(self::STATUS_NO_CONTENT);
-                    $this->logger->log(self::VALIDATION_FAILED, ['fields' => $updatedRecipe], Logger::ALERT);
+                    $this->response->setStatusCode(self::STATUS_VALIDATION_FAILED);
+                    $this->logger->log(self::VALIDATION_FAILED, ['fields' => $data], Logger::ALERT);
                 }
             } else {
-                $this->response->setStatusCode(self::STATUS_VALIDATION_FAILED);
-                $this->logger->log(self::VALIDATION_FAILED, ['fields' => $data], Logger::ALERT);
+                $this->response->setStatusCode(self::STATUS_NOT_FOUND);
             }
-
-        } else {
-            $this->response->setStatusCode(self::STATUS_NOT_FOUND);
         }
         return $this->response;
     }
@@ -289,7 +305,6 @@ class RecipesController extends AbstractController {
         } else {
             $this->response->setStatusCode(self::STATUS_VALIDATION_FAILED);
             $this->response->setContent(self::VALIDATION_INVALID_SEARCH_QUERY);
-
         }
         return $this->response;
     }
@@ -324,7 +339,6 @@ class RecipesController extends AbstractController {
         if (empty($responseJson)) {
             $this->response->setStatusCode(self::STATUS_NOT_FOUND);
             $this->response->setContent(self::VALIDATION_NO_RECORD);
-
             $this->logger->log(self::VALIDATION_NO_RECORD, ['id' => $recipeIdentifier], Logger::INFO);
         } else {
             $this->response->setContent($responseJson);
@@ -340,23 +354,20 @@ class RecipesController extends AbstractController {
      * @return Response
      * @throws Exception
      */
-    public function post(Request $request): Response {
+    public function post(Request $request, ValidatorInterface $validator, RecipesPostSchema $recipesPostSchema): Response {
         $pascalEm = (array)json_decode($request->getContent(), true);
-        $valid = $this->validateRequest($pascalEm);
+        $violations = $validator->validate($pascalEm, $recipesPostSchema::$schema);
+        $valid = $this->validateRequest($violations);
         if ($valid) {
-            $pascalEm = $this->normalizeData($pascalEm);
-            $validPostFields = $this->validateRecipeFields($pascalEm, 'POST');
-            $valid = false;
-            if ($pascalEm && is_array($pascalEm) && $validPostFields) {
-                $valid = $this->validateRequiredFields($pascalEm, __CLASS__.__FUNCTION__);
-            }
-            if ($valid) {
-                $recipeID = $this->recipesRepository->save($pascalEm);
+            $pascalEm = $this->normalizeRecipeData($pascalEm);
+            $recipeID = $this->recipesRepository->save($pascalEm);
+            if ($recipeID) {
                 $this->response->setStatusCode(self::STATUS_OK);
                 // Return posted data back.(use get to normalize ingredients array).
                 $this->getByIdInternal($recipeID);
             } else {
-                $this->response->setStatusCode(self::STATUS_VALIDATION_FAILED);
+                $this->response->setStatusCode(self::STATUS_EXCEPTION);
+                $this->response->setContent(self::VALIDATION_NO_RECORD);
             }
             $this->updateResponseHeader();
         }
@@ -369,12 +380,17 @@ class RecipesController extends AbstractController {
      * @param Request $request
      * @return bool $valid If request is valid.
      */
-    private function validateRequest($pascalEm): bool {
+    private function validateRequest($violations = []): bool {
         $valid = true;
-        if (empty($pascalEm)) {
+        if ($violations instanceof  ConstraintViolationListInterface && $violations->count() > 0) {
+            $formatedViolationList = [];
+            for ($i = 0; $i < $violations->count(); $i++) {
+                $violation = $violations->get($i);
+                $formatedViolationList[] = array($violation->getPropertyPath() => $violation->getMessage());
+            }
+            $this->logger->log('Schema validation failed', $formatedViolationList, Logger::CRITICAL);
+            $this->response->setContent('Schema validation failed');
             $this->response->setStatusCode(self::STATUS_VALIDATION_FAILED);
-            $this->response->setContent(self::VALIDATION_FAILED);
-            $this->logger->log('Invalid request object.', ['request content'=> $pascalEm], self::STATUS_VALIDATION_FAILED);
             $valid = false;
         }
         return $valid;
@@ -416,18 +432,19 @@ class RecipesController extends AbstractController {
      * @param array $parameters
      * @return array Normalized Data
      */
-    private function normalizeData(array $parameters): array {
+    private function normalizeRecipeData(array $parameters): array {
+        // TODO this shouldn't be needed with Schema validation.
         $normalizedData = [];
         foreach($parameters as $param => $value) {
-            if ($param === 'name' || $param === 'added_by') {
+            if ($param === 'name' || $param === 'addedBy') {
                 $value =  ucwords($value);
             }
             $normalizedData += [strtolower($param) => $value];
         }
         $normalizedData['favourites'] = $normalizedData['favourites'] ?? 0;
-        $normalizedData['added_by'] = $normalizedData['added_by'] ?? '';
-        $normalizedData['prep_time'] = $normalizedData['prep_time'] ?? null;
-        $normalizedData['cooking_time'] = $normalizedData['cooking_time'] ?? null;
+        $normalizedData['addedBy'] = $normalizedData['addedBy'] ?? '';
+        $normalizedData['prepTime'] = $normalizedData['prepTime'] ?? null;
+        $normalizedData['cookingTime'] = $normalizedData['cookingTime'] ?? null;
         $normalizedData['calories'] = $normalizedData['calories'] ?? null;
         $normalizedData['cuisine'] = $normalizedData['cuisine'] ?? '';
         $normalizedData['url'] = $normalizedData['url'] ?? '';
@@ -443,6 +460,7 @@ class RecipesController extends AbstractController {
      * @return bool
      */
     private function validateRequiredFields(array $parameters, $sender): bool {
+        // TODO This should be needed with Schema validation.
         $valid = isset($parameters['name']) && (isset($parameters['category']) && $this->validateCategory($parameters['category'])) && isset($parameters['directions']) && isset($parameters['ingredients']);
         if (!$valid) {
             $this->response->setContent(self::VALIDATION_FAILED);
@@ -460,6 +478,7 @@ class RecipesController extends AbstractController {
      * @return bool
      */
     private function validateCategory($category): bool {
+        // TODO this shoul be part of validation schema.
         $category = strtolower($category);
         return in_array($category, self::$categories);
     }
